requirements.txt
anthropic==0.26.0
python-binance==1.0.19
ccxt==4.3.0
pandas==2.2.0
numpy==1.26.0
chromadb==0.4.22
fastapi==0.110.0
uvicorn==0.29.0
python-telegram-bot==21.1.1
websockets==12.0
aiohttp==3.9.3
python-dotenv==1.0.1
yfinance==0.2.37
youtube-transcript-api==0.6.2
PyPDF2==3.0.1
httpx==0.27.0
pytest==8.1.0
pytest-asyncio==0.23.5
